+++ packages/cli/lib/_version.d.ts
@@ -1,1 +1,1 @@
-export declare const version = "cli/5.0.0-beta.147";
+export declare const version = "cli/5.0.0-beta.148";