+++ packages/ethers/lib.esm/_version.d.ts
@@ -1,1 +1,1 @@
-export declare const version = "ethers/5.0.0-beta.164";
+export declare const version = "ethers/5.0.0-beta.165";